refactor: introduce note editor component
