+++ src/features/NoteEditor/NoteEditor.tsx
@@ -1,0 +1,68 @@
+import React, { useCallback } from 'react';
+import { Box, HStack } from '@chakra-ui/react';
+import { useFilesRegistry } from '@features/App/Workspace/WorkspaceProvider';
+import { useAppSelector } from '@state/redux/hooks';
+import { selectEditorMode } from '@state/redux/settings/settings';
+
+import { FileUploader } from '../MonakoEditor/features/useDropFiles';
+import { MonacoEditor } from '../MonakoEditor/MonacoEditor';
+import { EditorPanelContext } from './EditorPanel';
+import { EditorPanel } from './EditorPanel/EditorPanel';
+import { RichEditor } from './RichEditor/RichEditor';
+
+export const NoteEditor = ({
+	text,
+	setText,
+}: {
+	text: string;
+	setText: (text: string) => void;
+}) => {
+	const editorMode = useAppSelector(selectEditorMode);
+
+	const filesRegistry = useFilesRegistry();
+	const uploadFile: FileUploader = useCallback(
+		async (file) => {
+			return filesRegistry.add(file);
+		},
+		[filesRegistry],
+	);
+
+	return (
+		<EditorPanelContext>
+			<HStack align="start" w="100%" overflowX="auto" flexShrink={0}>
+				<EditorPanel />
+			</HStack>
+
+			<HStack
+				sx={{
+					display: 'flex',
+					width: '100%',
+					height: '100%',
+					overflow: 'hidden',
+				}}
+			>
+				{(editorMode === 'plaintext' || editorMode === 'split-screen') && (
+					<Box
+						as={MonacoEditor}
+						value={text}
+						setValue={setText}
+						flexGrow="100"
+						uploadFile={uploadFile}
+						width="100%"
+						height="100%"
+						minW="0"
+					/>
+				)}
+				{(editorMode === 'richtext' || editorMode === 'split-screen') && (
+					<RichEditor
+						placeholder="Write your thoughts here..."
+						value={text}
+						onChanged={setText}
+					/>
+				)}
+			</HStack>
+		</EditorPanelContext>
+	);
+};
+
+NoteEditor.displayName = 'NoteEditor';
